[Refactor] Calendar date 처리 로직을 CalendarStore로 분리
